Moving around some functionality, Dashboard no longer requires spotify code to render
