+++ client/src/SignUpPage.jsx
@@ -17,8 +17,8 @@ const SignUpPage = () => {
   };
 
   return currentUser ? (
-    <div className="sign-in-page">
-      <div className="sign-in-title">You are logged in.</div>
+    <div>
+      <SpotifyLogin />
     </div>
   ) : (
     <div className="sign-in-page">
